Moved plot_qscore_heatmap logic to cpp (cherry picked from commit 3517ea8)
